Siehe Beschreibung für mehr.
- Eine Dokumentation zur Spieldaten Struktur wurde hinzugefügt.

- Würfel-Rolls Anzahl falls alle spielbaren Figuren im Haus sind kann nun zwischen 1 und 3 umgeschaltet werden.

- Die Spiel-Einstellungen lassen sich nun mit dem Steuerkreuz bedienen.

+++ 3ds/source/main.cpp
@@ -33,7 +33,7 @@ bool exiting = false;
 C2D_SpriteSheet sets, sprites;
 touchPosition touch;
 std::unique_ptr<Settings> konfiguration = nullptr;
-int fadeAlpha = 0;
+uint8_t fadeAlpha = 0;
 
 /*
 	Falls eine bestimmte position berührt wird.. mache etwas.
